Record four failed optimizations; default the expert bank to fp16
int8 mixers, int4 in_proj, a GPU router and a rank-4 mixer fold were all
tried on the 160 ms speculative pass and all cost more than they saved. The
first two disprove the weight-bandwidth model the layer was being tuned
against: halving in_proj, the largest weight in the layer, changed the time by
1% and the error by 8x. The mixer fold was bit-identical and 20% slower, so it
is not simply dispatch-bound either — the ANE wants work on the channel axis.

What did land, worth about 3 ms a pass: the recurrent state now lives in its
input IOSurface rather than being staged through a host array and copied
twice, only the one live column of the conv cache surface is written, and the
staging buffer is hoisted out of the layer loop.

The expert bank now defaults to fp16 scales and biases. That is 15 GB off a
76 GB bank, which is what keeps it resident once the drafter is loaded, and it
is also the faster pair. Plain decode is 111 ms/token and still matches BF16
greedy.

+++ runtime/flashnext_mlx_moe.py
@@ -17,7 +17,10 @@ class ResidentMoe:
     _configured = False
 
     def __init__(self, layer, shared, path=MLX4_DEFAULT):
-        self.dtype = getattr(mx, os.environ.get("FLASHNEXT_MOE_DTYPE", "float32"))
+        # fp16 throughout. fp32 scales and biases were 15 GB of a 76 GB bank,
+        # and on a 137 GB machine that margin decides whether the bank stays
+        # resident — see docs/W8A8-PROJECTIONS.md. It is also the faster pair.
+        self.dtype = getattr(mx, os.environ.get("FLASHNEXT_MOE_DTYPE", "float16"))
         # gather_qmm promotes FP16 + BF16 to FP32, then casts *all* scales
         # and biases before selecting experts (mlx/ops.cpp). With 48 banks
         # this creates ~15 GB/token of conversion output. Materialize the
@@ -69,6 +72,11 @@ class ResidentMoe:
         up = project(x, self.projections[1])
         y = project(gate * mx.sigmoid(gate) * up, self.projections[2])
         return mx.sum(y.squeeze(-2) * scores, axis=-2)
+
+    # A GPU router (softmax + argpartition + gather in one graph, one eval a
+    # layer) was tried and is a net loss: it removed the 6 ms of host routing
+    # and added 34 ms to the gather. The MoE eval is latency-bound, and the
+    # extra ops in the same graph cost more than the NumPy matmul they replace.
 
     def routed_multi(self, x_k, ids_k, scores_k):
         """Routed experts for k tokens in one submit.
